requirements.txt
streamlit
groq
langchain
langchain-groq
wikipedia
sentence-transformers
fpdf2
python-dotenv